+++ tests/server_event.rs
@@ -52,8 +52,6 @@ fn sending_receiving() {
         (SendMode::Broadcast, 1),
         (SendMode::Direct(SERVER_ID), 0),
         (SendMode::Direct(client_id), 1),
-        (SendMode::BroadcastExcept(SERVER_ID), 1),
-        (SendMode::BroadcastExcept(client_id), 0),
     ] {
         server_app
             .world
@@ -136,8 +134,6 @@ fn local_resending() {
         (SendMode::Broadcast, 1),
         (SendMode::Direct(SERVER_ID), 1),
         (SendMode::Direct(DUMMY_CLIENT_ID), 0),
-        (SendMode::BroadcastExcept(SERVER_ID), 0),
-        (SendMode::BroadcastExcept(DUMMY_CLIENT_ID), 1),
     ] {
         app.world
             .resource_mut::<Events<ToClients<DummyEvent>>>()